feat(dto): create UsuarioResponse and UsuarioUpdateRequest DTOs

+++ src/main/java/com/project/autobackend1/repository/UsuarioRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<usuario, Integer> {
     Optional<usuario> findByEmail(String email);
 }
